added trainee-toggle pipe and renamed files to match naming convention
