+++ src/app/User/user/Components/view-trainees/fonts/Futura-Std-Medium.otf
