+++ src/app/User/user/Pipes/trainee-toggle.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { TraineeTogglePipe } from './trainee-toggle.pipe';
+
+describe('TraineeTogglePipe', () => {
+  it('create an instance', () => {
+    const pipe = new TraineeTogglePipe();
+    expect(pipe).toBeTruthy();
+  });
+});
