+++ src/app/User/user/types/trainee-flag.ts
