+++ src/app/User/user/types/training-status.ts
