+++ src/app/User/user/user.module.ts
@@ -1,14 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ViewtraineesComponent } from './Components/viewtrainees/viewtrainees.component';
+import { ViewtraineesComponent } from './Components/view-trainees/view-trainees.component';
 import { FormsModule } from '@angular/forms';
+import { TraineeTogglePipe } from './Pipes/trainee-toggle.pipe';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule
   ],
-  declarations: [ViewtraineesComponent],
+  declarations: [ViewtraineesComponent, TraineeTogglePipe],
   exports: [
     ViewtraineesComponent
   ]
